// apps/fcm/gen/checkin.ts
export interface ChromeBuildProto {
  platform?: ChromeBuildProto_Platform | undefined;
  chromeVersion?: string | undefined;
  channel?: ChromeBuildProto_Channel | undefined;
}

export enum ChromeBuildProto_Platform {
  PLATFORM_WIN = 1,
  PLATFORM_MAC = 2,
  PLATFORM_LINUX = 3,
  PLATFORM_CROS = 4,
  PLATFORM_IOS = 5,
  PLATFORM_ANDROID = 6,
  UNRECOGNIZED = -1,
}
// ...
export enum ChromeBuildProto_Channel {
  CHANNEL_STABLE = 1,
  CHANNEL_BETA = 2,
  CHANNEL_DEV = 3,
  CHANNEL_CANARY = 4,
  CHANNEL_UNKNOWN = 5,
  UNRECOGNIZED = -1,
}
// ...
export interface AndroidCheckinProto {
  lastCheckinMsec?: string | undefined;
  cellOperator?: string | undefined;
  simOperator?: string | undefined;
  roaming?: string | undefined;
  userNumber?: number | undefined;
  type?: AndroidCheckinProto_DeviceType | undefined;
  chromeBuild?: ChromeBuildProto | undefined;
}

export enum AndroidCheckinProto_DeviceType {
  DEVICE_ANDROID_OS = 1,
  DEVICE_IOS_OS = 2,
  DEVICE_CHROME_BROWSER = 3,
  DEVICE_CHROME_OS = 4,
  UNRECOGNIZED = -1,
}
// ...
export interface GservicesSetting {
  name?: Uint8Array | undefined;
  value?: Uint8Array | undefined;
}

export interface AndroidCheckinRequest {
  imei?: string | undefined;
  id?: string | undefined;
  digest?: string | undefined;
  checkin?: AndroidCheckinProto | undefined;
  desiredBuild?: string | undefined;
  locale?: string | undefined;
  loggingId?: string | undefined;
  marketCheckin?: string | undefined;
  macAddr?: string[] | undefined;
  meid?: string | undefined;
  accountCookie?: string[] | undefined;
  timeZone?: string | undefined;
  securityToken?: string | undefined;
  version?: number | undefined;
  otaCert?: string[] | undefined;
  serialNumber?: string | undefined;
  esn?: string | undefined;
  macAddrType?: string[] | undefined;
  fragment?: number | undefined;
  userName?: string | undefined;
  userSerialNumber?: number | undefined;
}

export interface AndroidCheckinResponse {
  statsOk?: boolean | undefined;
  timeMsec?: string | undefined;
  digest?: string | undefined;
  settingsDiff?: boolean | undefined;
  deleteSetting?: string[] | undefined;
  setting?: GservicesSetting[] | undefined;
  marketOk?: boolean | undefined;
  androidId?: string | undefined;
  securityToken?: string | undefined;
  versionInfo?: string | undefined;
}

function createBaseChromeBuildProto(): ChromeBuildProto {
  return { platform: 1, chromeVersion: "", channel: 1 };
}

export const ChromeBuildProto: MessageFns<ChromeBuildProto> = {
  encode(message: ChromeBuildProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.platform !== undefined) {
      writer.uint32(8).int32(message.platform);
    }
    if (message.chromeVersion !== undefined) {
      writer.uint32(18).string(message.chromeVersion);
    }
    if (message.channel !== undefined) {
      writer.uint32(24).int32(message.channel);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ChromeBuildProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseChromeBuildProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.platform = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.chromeVersion = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.channel = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ChromeBuildProto {
    return {
      platform: isSet(object.platform) ? chromeBuildProto_PlatformFromJSON(object.platform) : 1,
      chromeVersion: isSet(object.chromeVersion)
        ? globalThis.String(object.chromeVersion)
        : isSet(object.chrome_version)
        ? globalThis.String(object.chrome_version)
        : "",
      channel: isSet(object.channel) ? chromeBuildProto_ChannelFromJSON(object.channel) : 1,
    };
  },

  toJSON(message: ChromeBuildProto): unknown {
    const obj: any = {};
    if (message.platform !== undefined && message.platform !== 1) {
      obj.platform = chromeBuildProto_PlatformToJSON(message.platform);
    }
    if (message.chromeVersion !== undefined && message.chromeVersion !== "") {
      obj.chromeVersion = message.chromeVersion;
    }
    if (message.channel !== undefined && message.channel !== 1) {
      obj.channel = chromeBuildProto_ChannelToJSON(message.channel);
    }
    return obj;
  },

  create(base?: DeepPartial<ChromeBuildProto>): ChromeBuildProto {
    return ChromeBuildProto.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ChromeBuildProto>): ChromeBuildProto {
    const message = createBaseChromeBuildProto();
    message.platform = object.platform ?? 1;
    message.chromeVersion = object.chromeVersion ?? "";
    message.channel = object.channel ?? 1;
    return message;
  },
};
// ...
export const GservicesSetting: MessageFns<GservicesSetting> = {
  encode(message: GservicesSetting, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== undefined && message.name.length !== 0) {
      writer.uint32(10).bytes(message.name);
    }
    if (message.value !== undefined && message.value.length !== 0) {
      writer.uint32(18).bytes(message.value);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GservicesSetting {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGservicesSetting();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.bytes();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.value = reader.bytes();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GservicesSetting {
    return {
      name: isSet(object.name) ? bytesFromBase64(object.name) : new Uint8Array(0),
      value: isSet(object.value) ? bytesFromBase64(object.value) : new Uint8Array(0),
    };
  },

  toJSON(message: GservicesSetting): unknown {
    const obj: any = {};
    if (message.name !== undefined && message.name.length !== 0) {
      obj.name = base64FromBytes(message.name);
    }
    if (message.value !== undefined && message.value.length !== 0) {
      obj.value = base64FromBytes(message.value);
    }
    return obj;
  },

  create(base?: DeepPartial<GservicesSetting>): GservicesSetting {
    return GservicesSetting.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GservicesSetting>): GservicesSetting {
    const message = createBaseGservicesSetting();
    message.name = object.name ?? new Uint8Array(0);
    message.value = object.value ?? new Uint8Array(0);
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}